pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metalabel-project"
version = "0.1.0"
description = "Practical quant finance repository for a primary model with future meta-labeling expansion."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "openpyxl>=3.1",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "scikit-learn>=1.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
metalabel = "metalabel.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
